+++ index.html
@@ -30,7 +30,9 @@
     </div>
     </div>
   </body>
-  <script src="./main.js"></script>
-<script src="./api.js"></script>
-<script src="./renderStudents.js"></script>
+  <script type="module" src="./utils.js"></script>
+  <script type="module" src="./main.js"></script>
+<script type="module" src="./api.js"></script>
+<script type="module" src="./renderComments.js"></script>
+
 </html>
