+++ renderComments.js
@@ -1,0 +1,51 @@
+const getListCommentsEdit = (comment, index) => {
+  let likeClass = comment.active ? "-active-like" : "";
+     return `<li class="comment">
+           <div class="comment-header">
+             <div id ="name">${comment.name}</div>
+             <div>${comment.date} </div>
+           </div>
+           <div class="comment-body">
+             <div id = "text" data-index="${index}" class="comment-text" style={{ whiteSpace: 'pre-line' }}>
+               ${comment.text}
+             </div>
+           </div>
+           <div class="comment-footer">
+             <button data-index="${index}" class="edited-button">Редактировать</button>
+             <div class="likes">
+               <span class="likes-counter">${comment.like}</span> ${comment.like === 1 ? 'лайк' : 'лайков'}
+               <button data-index="${index}" class="like-button ${likeClass}"></button> 
+             </div>
+           </div>
+         </li>`
+}
+
+export const renderComments = (comments, handleCommentEditClick, handleCommentFeedbackClick, handleCommentLikeClick) => {
+    const commentsHtml = comments.map((comment, index) => getListCommentsEdit(comment, index))
+     .join('');
+const listElement = document.getElementById("list");
+     listElement.innerHTML = commentsHtml;
+    setTimeout(() => {
+      onCommentLikeClick(handleCommentLikeClick);
+      onCommentEditClick(handleCommentEditClick);
+      onCommentFeedbackClick(handleCommentFeedbackClick)
+    }, 1000) 
+   }
+export const onCommentLikeClick = (handleCommentLikeClick) => {
+    const likeButtonsElements = document.querySelectorAll(".like-button");
+    for (const likeButtonsElement of likeButtonsElements) {
+      likeButtonsElement.addEventListener("click", handleCommentLikeClick);
+    }
+  };
+export const onCommentEditClick = (handleCommentEditClick) => { 
+    const editButtonsElements = document.querySelectorAll(".edited-button"); 
+    for (const editButtonElement of editButtonsElements) { 
+      editButtonElement.addEventListener("click", handleCommentEditClick);
+    } 
+  };
+export const onCommentFeedbackClick = (handleCommentFeedbackClick) => {  
+    const feedbackButtonsElements = document.querySelectorAll(".comment-text");   
+    for (const feedbackButtonsElement of feedbackButtonsElements) {   
+      feedbackButtonsElement.addEventListener("click", handleCommentFeedbackClick);  
+    }  
+  };
